scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "ruff>=0.9.0",
]
